+++ CHANGELOG.md
@@ -38,6 +38,7 @@ var value = from {
 ```
 - `recursive_call_limit` build option limit recursive calls (default to 200)
 - Compiler will warn about code after a `return` statement
+- Compiler will warn about unreferenced imports (https://github.com/buzz-language/buzz/issues/272)
 - `namespace` (https://github.com/buzz-language/buzz/issues/271): if a script exports at least one symbol, it has to define a namespace for the script with `namespace mynamespace`
     - By default, imported symbols from another script will be under `libprefix.XXXX`
     - When importing something, you can still redefine its namespace prefix with `import "..." as mynewnamespace` or remove it altogether with `import "..." _`